requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.28.1
SQLAlchemy==1.4.32
pymysql==1.0.2
gunicorn==20.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==1.10.11
python-multipart==0.0.6
celery==5.3.1
redis==4.6.0
elasticsearch==7.17.9
httpx==0.24.1
pytest==7.4.0
jsonsearch==0.0.2